Proof of Concept DB connection, adding querying etc, Event List Component Creation
Created a simple docker container system to run a postgresql server locally, added seeding support for sample testing data, and confirmation of writing to a table and reading from a table, as seen in the new component event-list.tsx.

Commented out the old mongoDB DB connection and replaced with postgres drizzle
Added detailed notes for documentation on database setup and structure.