Applied Documentation Processing to Database System
Created various documentation markdown files for the schema references, maintenance and visualization. As well as a few other small changes for preparing. You can see an example of useage via events page by going to /dashboard, clicking on Create Event, and thus it should show up in the Events list. Note, that is for demonstration purposes. No other API route is built at the moment for database connections.